desktop/requirements.txt
python-can[gs_usb]
protobuf==4.21.6